Add Same Site Cookie Support
This adds Same Site Cookie Support, a new security feature supported by Chrome 51 (to be released in May to the stable channel) and Firefox has intentions to add it in Q2 2016 as well.

SameSite Cookies allows to specify from which origins a cookie is sent. The policy of "lax" prevents the cookie from being sent to ownCloud from any other domain than the one that ownCloud is running on. Effectively this prevents CSRF attacks on any HTTP method other than GET which is a very important and big security hardening.

This implementation behaves as following:

1. For browsers: If the cookie is not sent AND cookies are supported then add the cookie and reload the page.
2. For DAV clients: If the cookie is not sent AND cookies are supported then add the cookie and issue a 503 status code.

The implementation has the advantage that it is not intrusive and does not break existing clients or pages.